requirements.txt
beautifulsoup4==4.7.1
bottle==0.12.16
feedparser==5.2.1
lxml==4.3.1
Mastodon.py==1.3.1
mysqlclient==1.4.2.post1
oauthlib==2.0.6
psycopg2-binary==2.7.7
python-dotenv==0.10.1
requests==2.21.0
requests-unixsocket==0.1.5
SQLAlchemy==1.2.18
urllib3==1.24.1
